Issue #17 Moved lexis from under util to first class
