+++ code/lexis/LexisCalculator.py
